# store/api/docs/swagger.yaml
definitions:
...
  models.CreateProduct:
    properties:
      category_id:
        type: string
      name:
        type: string
      original_price:
        type: integer
      price:
        type: integer
      quantity:
        type: integer
    type: object
...
  models.Product:
    properties:
      category_id:
        type: string
      id:
        type: string
      name:
        type: string
      original_price:
        type: integer
      price:
        type: integer
      quantity:
        type: integer
    type: object
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.SellRequest:
    properties:
      basket_id:
        type: string
      products:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  models.UpdateProduct:
    properties:
      category_id:
        type: string
      id:
        type: string
      name:
        type: string
      original_price:
        type: integer
      price:
        type: integer
      quantity:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /product:
    post:
      consumes:
      - application/json
      description: create a new product
      parameters:
      - description: product
        in: body
        name: product
        schema:
          $ref: '#/definitions/models.CreateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create a new product
      tags:
      - product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: delete product
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete product
      tags:
      - product
    get:
      consumes:
      - application/json
      description: get product by id
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product by id
      tags:
      - product
    put:
      consumes:
      - application/json
      description: update product
      parameters:
      - description: product_id
        in: path
        name: id
        required: true
        type: string
      - description: product
        in: body
        name: product
        schema:
          $ref: '#/definitions/models.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update product
      tags:
      - product
  /products:
    get:
      consumes:
      - application/json
      description: get product list
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get product list
      tags:
      - product
  /sell-new:
    post:
      consumes:
      - application/json
      description: sell product
      parameters:
      - description: sell
        in: body
        name: sell
        schema:
          $ref: '#/definitions/models.SellRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: sell product
      tags:
      - sell
...
swagger: "2.0"